Reapply "Split NAPI modules [PM-31598] (#18722)" (#19112)
This reverts commit cae1ae649162c71a87efe388cce24d2419a730f9.
